+++ packages/manager/apps/dedicated/client/app/dedicatedCloud/datacenter/backup/backup.routing.js
@@ -104,13 +104,11 @@ export default /* @ngInject */ ($stateProvider) => {
         dedicatedCloudDatacenterBackupService,
         productId,
         datacenterId,
-        currentUser,
         backup,
       ) =>
         dedicatedCloudDatacenterBackupService.getBackupOffers(
           productId,
           datacenterId,
-          currentUser.ovhSubsidiary,
           backup,
         ),
       backupConditionsUrl: /* @ngInject */ (currentUser) =>