+++ packages/manager/apps/hub/src/billing/types/service-links.type.ts
@@ -1,0 +1,18 @@
+export type ServiceLinkName =
+  | 'anticipatePayment'
+  | 'buySMSCredits'
+  | 'cancelCommitment'
+  | 'cancelResiliation'
+  | 'configureExchangeAccountsRenewal'
+  | 'configureRenewal'
+  | 'configureSMSAutoReload'
+  | 'manageCommitment'
+  | 'modifyExchangeBilling'
+  | 'payBill'
+  | 'renewManually'
+  | 'resiliate'
+  | 'resiliateByDeletion'
+  | 'seeService'
+  | 'warnBillingNic';
+
+export type ServiceLinks = Partial<Record<ServiceLinkName, string>>;