+++ packages/manager/modules/nutanix/src/dashboard/nodes/list/index.js
@@ -7,7 +7,7 @@ const moduleName = 'ovhManagerNutanixNodesAllLazyLoading';
 angular.module(moduleName, ['ui.router', 'oc.lazyLoad']).config(
   /* @ngInject */ ($stateProvider) => {
     $stateProvider.state('nutanix.dashboard.nodes.all.**', {
-      url: '',
+      url: '/',
       lazyLoad: ($transition$) => {
         const $ocLazyLoad = $transition$.injector().get('$ocLazyLoad');