+++ packages/manager/modules/nutanix/src/dashboard/nodes/list/module.js
@@ -1,5 +1,4 @@
 import angular from 'angular';
-import '@ovh-ux/manager-core';
 import '@uirouter/angularjs';
 import 'angular-translate';
 import { region, iamProtectedData } from '@ovh-ux/manager-components';
@@ -7,17 +6,36 @@ import { region, iamProtectedData } from '@ovh-ux/manager-components';
 import routing from './routing';
 import component from './component';
 import datacenterName from '../../component/datacenter-name/module';
+import nodeStatus from '../../component/node-os-datagrid/module';
+import nodeOsDatagrid from '../../component/service-status/module';
+import nodeActionMenu from '../../component/node-action-menu/module';
+import nodeActionMenuTerminated from '../../component/node-action-menu-terminated/module';
+import addNodes from './add-nodes';
+import poweroffNode from './poweroff-node';
+import installNode from './install-node';
+import reinstallNode from './reinstall-node';
+import uninstallNode from './uninstall-node';
+import resiliateNode from './resiliate-node';
 
 const moduleName = 'ovhManagerNutanixAllNodes';
 
 angular
   .module(moduleName, [
-    'ovhManagerCore',
     'pascalprecht.translate',
     'ui.router',
     region,
     datacenterName,
     iamProtectedData,
+    nodeStatus,
+    nodeOsDatagrid,
+    nodeActionMenu,
+    nodeActionMenuTerminated,
+    addNodes,
+    poweroffNode,
+    installNode,
+    reinstallNode,
+    uninstallNode,
+    resiliateNode,
   ])
   .config(routing)
   .component('nutanixAllNodes', component)